chThread.h: unconditionally use libc11 threads.h
Whose idiotic idea was it to require compilers to define a macro when a feature
ISN'T supported? I could compile this with something ancient like gcc 2.95 and
it would believe that since __STDC_NO_THREADS__ isn't defined that I have
a compiler that supports C11 threads. And before you say "check the __STDC__
macro value", that doesn't help because there are versions of GCC that claim
C11 support but also don't define __STDC_NO_THREADS__ (even though the feature
is not supported).

Signed-off-by: Steven Noonan <steven@uplinklabs.net>