reorganize how algorithm iteration is handled
Now it's a bit more easy to manage. Instead of all sorts of tapdancing around
selecting values from an enum, it's just a matter of selecting items in an
array. Items that aren't in the array aren't included in the algorithm
iteration. Easy!

I've also disabled some GPU algorithms that weren't behaving (based on
crosscheck results).

Signed-off-by: Steven Noonan <steven@uplinklabs.net>