actions: make compatible with 'act'
'act' is from https://github.com/nektos/act

The first problem is that act tries to run the workflows in parallel on
the same host, but they end up all using the same docker container name.
Set the docker container based on the build matrix, and use that same
name throughout the whole run of each workflow.

The second problem is that we were bind-mounting the source tree, which
causes problems since act runs multiple matrix combinations in parallel
on the same container.

The final issue is that we can't assume the environment is perfectly
clean if we've run act once already, so we need to destroy any old
docker container that may be running, and destroy the container at the
end of a successful run.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>